Move vertical-nav components into this directory
