--- components/atoms/nav-footer/styles.css
@@ -1,11 +1,0 @@
-.navFooter {
-  height: 60px;
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  font-size: 12px;
-  color: var(--color-secondary);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
