--- components/atoms/navigation-item/styles.css
@@ -1,24 +1,0 @@
-.navItem {
-  font-size: 14px;
-  line-height: 16.1px;
-  padding: 20px 24px;
-  text-align: left;
-  font-family: 'Source Sans Pro', sans-serif;
-
-  &:hover {
-    background-color: var(--color-primary-dark);
-  }
-}
-
-.link {
-  display: flex;
-  align-items: center;
-  text-decoration: none;
-  color: white;
-}
-
-.icon {
-  margin-right: 15px;
-  width: 20px;
-  height: 20px;
-}
