--- components/molecules/navigation-items/styles.css
@@ -1,5 +1,0 @@
-.navItems {
-  list-style-type: none;
-  margin-top: 0;
-  padding-left: 0;
-}
