+++ components/organisms/vertical-nav/footer.js
