+++ components/organisms/vertical-nav/index.stories.js
@@ -4,8 +4,6 @@ import { action } from '@storybook/addon-actions'
 import { withInfo } from '@storybook/addon-info'
 import { withKnobs } from '@storybook/addon-knobs/react'
 
-import NavigationItems from '_components/molecules/navigation-items'
-import NavigationItem from '_components/atoms/navigation-item'
 import VerticalNav from '_components/organisms/vertical-nav'
 import settingsIcon from '_components/assets/icons/settings.png'
 import cklLogo from '_components/assets/icons/ckl-logo.png'
@@ -16,23 +14,23 @@ storiesOf('Organisms/VerticalNav', module)
     'vertical nav',
     withInfo({ text: '' })(() => (
       <VerticalNav navFooterIcon={cklLogo} navFooterText="2018 Todos os direitos reservados">
-        <NavigationItems>
-          <NavigationItem link="teste" iconSrc={settingsIcon}>
+        <VerticalNav.Items>
+          <VerticalNav.Item link="teste" iconSrc={settingsIcon}>
             Menu 1
-          </NavigationItem>
-          <NavigationItem link="teste" iconSrc={settingsIcon}>
+          </VerticalNav.Item>
+          <VerticalNav.Item link="teste" iconSrc={settingsIcon}>
             Menu 2
-          </NavigationItem>
-          <NavigationItem link="teste" iconSrc={settingsIcon}>
+          </VerticalNav.Item>
+          <VerticalNav.Item link="teste" iconSrc={settingsIcon}>
             Menu 3
-          </NavigationItem>
-          <NavigationItem link="teste" iconSrc={settingsIcon}>
+          </VerticalNav.Item>
+          <VerticalNav.Item link="teste" iconSrc={settingsIcon}>
             Menu 4
-          </NavigationItem>
-          <NavigationItem link="teste" iconSrc={settingsIcon}>
+          </VerticalNav.Item>
+          <VerticalNav.Item link="teste" iconSrc={settingsIcon}>
             Menu 5
-          </NavigationItem>
-        </NavigationItems>
+          </VerticalNav.Item>
+        </VerticalNav.Items>
       </VerticalNav>
     ))
   )
