+++ components/organisms/vertical-nav/item.js
@@ -3,12 +3,11 @@ import PropTypes from 'prop-types'
 import classNames from 'classnames'
 
 import Icon from '_components/atoms/icon'
-import drawerStyle from '_components/organisms/drawer/styles.css'
 
 import styles from './styles.css'
 
 const NavigationItem = ({ iconSrc, link, className, children }) => (
-  <li className={classNames(styles.navItem, drawerStyle.navItem)}>
+  <li className={classNames(styles.navItem, styles.navItem)}>
     <a className={styles.link} href={link}>
       <Icon className={styles.icon} src={iconSrc} />
       <span className={className}>{children}</span>
