+++ components/organisms/vertical-nav/styles.css
@@ -80,3 +80,46 @@
   height: 60px;
   width: 60px;
 }
+
+.navItems {
+  list-style-type: none;
+  margin-top: 0;
+  padding-left: 0;
+}
+
+.navItem {
+  font-size: 14px;
+  line-height: 16.1px;
+  padding: 20px 24px;
+  text-align: left;
+  font-family: 'Source Sans Pro', sans-serif;
+
+  &:hover {
+    background-color: var(--color-primary-dark);
+  }
+}
+
+.link {
+  display: flex;
+  align-items: center;
+  text-decoration: none;
+  color: white;
+}
+
+.icon {
+  margin-right: 15px;
+  width: 20px;
+  height: 20px;
+}
+
+.navFooter {
+  height: 60px;
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  font-size: 12px;
+  color: var(--color-secondary);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
